refactor: move deprecated js-dash-sdk files to dedicated subdirectories
