+++ test/js-dash-sdk-deprecated/goodNodes.js
