+++ tutorials/js-dash-sdk-deprecated/checkNetworkConnection.js
