+++ tutorials/js-dash-sdk-deprecated/contract/registerContractProvided.js
