+++ tutorials/js-dash-sdk-deprecated/document/deleteDocument.js
