+++ tutorials/js-dash-sdk-deprecated/document/getDocument.js
