+++ tutorials/js-dash-sdk-deprecated/document/submitDocument.js
