+++ tutorials/js-dash-sdk-deprecated/getNewWalletInfo.js
