+++ tutorials/js-dash-sdk-deprecated/identity/registerIdentity.js
