+++ tutorials/js-dash-sdk-deprecated/name/registerAlias.js
