+++ tutorials/js-dash-sdk-deprecated/name/registerName.js
