+++ tutorials/js-dash-sdk-deprecated/name/retrieveNameByName.js
