+++ tutorials/js-dash-sdk-deprecated/name/retrieveNameByRecord.js
