+++ tutorials/js-dash-sdk-deprecated/name/retrieveNameBySearch.js
